requirements.txt
numpy==1.22.2
Pillow==9.4.0
PyPDF2==3.0.1
reportlab==3.6.11
tqdm==4.64.1